+++ examples/organization/src/components/organization/CreateOrganization.tsx
@@ -1,10 +1,14 @@
 import { Group, ID } from "jazz-tools";
 import { useState } from "react";
 import { useNavigate } from "react-router";
-import { Errors } from "./Errors.tsx";
+import { useAccount, useCoState } from "../../main.tsx";
+import {
+  DraftOrganization,
+  ListOfProjects,
+  Organization,
+} from "../../schema.ts";
+import { Errors } from "../Errors.tsx";
 import { OrganizationForm } from "./OrganizationForm.tsx";
-import { useAccount, useCoState } from "./main.tsx";
-import { DraftOrganization, ListOfProjects, Organization } from "./schema.ts";
 
 export function CreateOrganization() {
   const { me } = useAccount({
