+++ examples/organization/src/main.tsx
@@ -2,7 +2,10 @@ import { DemoAuthBasicUI, createJazzReactApp, useDemoAuth } from "jazz-react";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import Router from "./Router.tsx";
+import { RouterProvider, createHashRouter } from "react-router-dom";
+import { AcceptInvitePage } from "./AcceptInvitePage.tsx";
+import { HomePage } from "./HomePage.tsx";
+import { OrganizationPage } from "./OrganizationPage.tsx";
 import { JazzAccount } from "./schema.ts";
 
 const Jazz = createJazzReactApp({
@@ -10,6 +13,24 @@ const Jazz = createJazzReactApp({
 });
 
 export const { useAccount, useCoState, useAcceptInvite } = Jazz;
+function Router() {
+  const router = createHashRouter([
+    {
+      path: "/",
+      element: <HomePage />,
+    },
+    {
+      path: "/organizations/:organizationId",
+      element: <OrganizationPage />,
+    },
+    {
+      path: "/invite/*",
+      element: <AcceptInvitePage />,
+    },
+  ]);
+
+  return <RouterProvider router={router}></RouterProvider>;
+}
 
 function JazzAndAuth({ children }: { children: React.ReactNode }) {
   const [auth, authState] = useDemoAuth();
